Move all public EDU structs and enums to `Types.hpp/ipp`
