+++ Sts1CobcSw/Edu/CMakeLists.txt
@@ -1,4 +1,4 @@
-target_sources(Sts1CobcSw_Edu PRIVATE ProgramQueue.cpp ProgramStatusHistory.cpp)
+target_sources(Sts1CobcSw_Edu PRIVATE ProgramStatusHistory.cpp)
 target_link_libraries(
     Sts1CobcSw_Edu PUBLIC rodos::without_main_on_linux etl::etl Sts1CobcSw_FramSections
                           Sts1CobcSw_Outcome Sts1CobcSw_ProgramId Sts1CobcSw_Serial
